requirements.txt
arabic-reshaper==2.1.3
colorama==0.4.5
et-xmlfile==1.1.0
fpdf==1.7.2
future==0.18.2
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
Pillow==9.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
qrcode==7.3.1
six==1.16.0
